Split up webview content utils (DH-17600)
