+++ src/crossModule/webViewContentUtils/getIframeContentWindow.ts
@@ -1,0 +1,25 @@
+/**
+ * Code in this module needs to be consumable from both the extension code (CJS)
+ * and the webview content code (ESM). Avoid importing anything here from outside
+ * of the `crossModule` folder to minimize the risk of breaking the builds.
+ */
+
+import { assertDefined } from '../assertUtil';
+import { CONTENT_IFRAME_ID } from '../constants';
+
+/**
+ * Get the `contentWindow` of the content iframe in the WebView.
+ */
+export function getIframeContentWindow(): Window {
+  const maybeIframeEl = document.getElementById(CONTENT_IFRAME_ID);
+
+  assertDefined(maybeIframeEl, `Iframe with id ${CONTENT_IFRAME_ID} not found`);
+
+  if (!(maybeIframeEl instanceof HTMLIFrameElement)) {
+    throw new Error('Element is not an iframe');
+  }
+
+  assertDefined(maybeIframeEl.contentWindow, 'iframe content window');
+
+  return maybeIframeEl.contentWindow;
+}
